Factorio 2.0 Update Attempt
My best faith attempt to update the mod for 2.0.  It boots, you can get in-game, it even almost functions.

Items don't seem to be taken out of Provider chests, and while you can shove things into the network by deleting them, after doing so the on_nth_tick event will throw a "table index is nil" error against line 571 in GlobalState.lua that's had me stumped for several hours now.

I'm probably reaching the end of my knowledge here, and I don't know how useful any of it may be, but I figured I'd throw it up in case anyone else might see it and be able to do something with it, or it might be of some small use to the original author.

Hurling yourself face-first into the brick wall that is somebody else's code is a nightmare of a way to learn a programming language, that's for sure!